pyproject.toml
[project]
name = "todoist-mcp-server"
version = "1.0.0"
description = "MCP server for Todoist task management"
authors = [{ name = "Tomer Rosenfeld", email = "mail@tomerrosenfeld.com" }]
dependencies = [
    "todoist-api-python>=2.1.3",
    "tinydb>=4.8.0",
    "mcp>=0.9.1",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.30.1",
    "starlette>=0.37.2",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]
requires-python = ">=3.9"
readme = "README.md"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "todoist_api_python.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tinydb.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mcp.*"
ignore_missing_imports = true
